requirements.txt
# BioTransLM 项目依赖清单
# Python 版本要求: >=3.10, <3.12

# 核心深度学习框架
torch>=2.0.0,<3.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-sparse>=0.6.17  # 用于稀疏张量操作

# Transformers和NLP相关
transformers>=4.30.0,<5.0.0
tokenizers>=0.13.0
datasets>=2.12.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

# 向量检索库
faiss-cpu>=1.7.4; sys_platform != 'linux' or platform_machine != 'x86_64'
faiss-gpu>=1.7.4; sys_platform == 'linux' and platform_machine == 'x86_64'

# 科学计算和数据处理
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.2.2
matplotlib>=3.7.0
seaborn>=0.12.0

# 稀疏矩阵和图计算
networkx>=3.1
pyg-lib>=0.2.0
torch_geometric>=2.3.0

# 日志和配置
pyyaml>=6.0
structlog>=23.1.0
python-json-logger>=2.0.7
toml>=0.10.2
tomli>=2.0.1; python_version < '3.11'

# 测试和质量保证
pytest>=7.3.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
hypothesis>=6.75.0
mypy>=1.3.0
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0

# 性能分析和调试
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.0.2
debugpy>=1.6.7

# 工具和辅助
tqdm>=4.65.0
click>=8.1.3
rich>=13.3.5
typer>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# 分布式训练（可选）
# accelerate>=0.20.0
# deepspeed>=0.9.5

# 可视化
tensorboard>=2.13.0
# wandb>=0.15.0  # 可选，用于实验追踪

# 消息队列和序列化（如果需要）
# redis>=4.5.5
# msgpack>=1.0.5

# Jupyter和开发工具
notebook>=6.5.0
ipython>=8.13.0
ipywidgets>=8.0.6

# 设置工具
setuptools>=67.7.0
wheel>=0.40.0
pip>=23.1.0

# 质量检查工具
pre-commit>=3.3.0
bandit>=1.7.5
safety>=2.3.5
